generic_utilities/jira_testrail/requirements.txt
jira
pyyaml
atlassian-python-api


